Reorganized the DSL files a little bit.
